rename files but add lowercase imports
